create_ir_statement 함수 노출 및 모듈 노출
