+++ fireball/src/arch/mod.rs
@@ -1,3 +1,3 @@
 //! 여러 아키텍처 별 구현이 담겨있는 모듈
 
-mod x86_64;
+pub mod x86_64;
